+++ lib/src/storage/v3/package_api_storage_v3.g.dart
@@ -72,7 +72,6 @@ Map<String, dynamic> _$$_PackageApiStorageV3ToJson(
     };
 
 const _$PackageApiSemanticsEnumMap = {
-  PackageApiSemantics.mergeBaseClasses: 'mergeBaseClasses',
   PackageApiSemantics.containsPlatformConstraints:
       'containsPlatformConstraints',
 };